jstut.json driven overview process progress
We will now try and crawl all the source files in the package.  For
deuxdrop this results in a lot of errors being logged, but things
may actually be working.

The next step is to hook up a very primitive UI to try and display
something useful for the overview, probably a list of the modules
with the summary lines for the module overview comments.